auth module api
