+++ client/src/main/java/lx/talx/client/utils/Log.java
@@ -78,5 +78,6 @@ public class Log {
   }
 
   public static void log(String message) {
+    info(message);
   }
 }
